backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
asyncpg>=0.29.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
email-validator>=2.1.0
aiohttp>=3.9.0

# Testing dependencies
pytest-cov==4.1.0
